requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy
opencv-python
tqdm
tensorboard
tensorboardX
scikit-image
medpy
pillow
scipy
